+++ src/main/java/com/pplip/domain/board/notice/api/request/NoticeCommentRequest.java
@@ -5,14 +5,13 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-public class NoticeRequest {
+public class NoticeCommentRequest {
 
     @Data
     @NoArgsConstructor
     @AllArgsConstructor
     @Builder
     public static class Post {
-        private String title;
         private String content;
     }
 
@@ -21,7 +20,6 @@ public class NoticeRequest {
     @AllArgsConstructor
     @Builder
     public static class Update {
-        private String title;
         private String content;
     }
 }
